Support sub-commands on main binary
In order to ship a single binary which carries all the functionality we
want to be able to plug different comamnds into it. This will ease
opration and as well distribution of Configsum as system as well as
a tool.

* breakout config in own sub command
* support sub commands in main
* support global flag handling command flag handling
